chore: Refactor email layout and selection component from layout
