+++ src/app/(emails)/emails/layout.tsx
@@ -6,20 +6,8 @@ import Image from "next/image"
 import { notFound } from "next/navigation";
 import Button from "@/components/ui/button";
 import Selection from "@/components/selection";
-
-async function fetchEmailLimit() {
-  const response = await fetch(`${process.env.NEXT_PUBLIC_API_BASE_URL}/api/limit`, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
-  if (!response.ok) {
-    throw new Error('Failed to fetch the email limit');
-  }
-  const data = await response.json();
-  return data.limit;
-}
+import EmailList from "@/components/emailList";
+import fetchEmailLimit from "@/helpers/fetchlimit";
 
 const Layout = async ({ children }: { children: React.ReactNode }) => {
   const session = await getServerSession(authOptions);
@@ -68,33 +56,7 @@ const Layout = async ({ children }: { children: React.ReactNode }) => {
               <SignOutButton className="h-full aspect-square" />
             </div>
           </div>
-          <nav className="flex flex-1 flex-col overflow-auto max-h-[70vh]">
-            <ul role="list" className="flex flex-1 flex-col gap-y-4">
-              {emails.map((email: any) => {
-                const isUnread = email.labelIds.includes('UNREAD');
-                const isInInbox = email.labelIds.includes('INBOX');
-                const filteredLabels = email.labelIds.filter((label: string) => label === 'INBOX');
-    
-                return (
-                  <li key={email.id} className={`p-4 rounded-lg ${isUnread ? 'bg-blue-600' : 'bg-gray-700'} hover:bg-gray-600 transition-colors text-white`}>
-                    <div className="flex justify-between items-center mb-2">
-                      <span className={`font-semibold ${isUnread ? 'text-yellow-300' : 'text-gray-300'}`}>
-                        {isUnread ? 'Unread' : 'Read'}
-                      </span>
-                      {isInInbox && (
-                        <span className="bg-indigo-400 text-white text-xs px-2 py-1 rounded-full">
-                          {filteredLabels.join(', ')}
-                        </span>
-                      )}
-                    </div>
-                    <div className="mt-2 text-sm">
-                      {email.snippet}
-                    </div>
-                  </li>
-                );
-              })}
-            </ul>
-          </nav>
+          <EmailList emails={emails} />
         </div>
       </div>
     );
