+++ src/components/selection.tsx
@@ -4,6 +4,7 @@ import React, { useState } from 'react';
 const Selection = () => {
   const [selectedValue, setSelectedValue] = useState('5'); // Initialize with default value
 
+  
   const handleSelectChange = async (e: React.ChangeEvent<HTMLSelectElement>) => {
     console.log(e.target.value); // Log selected value
     setSelectedValue(e.target.value); // Update selected value
